# tides-server/swagger/swagger.yml
swagger: "2.0"
info:
  title: 'CloudTides-Server'
  version: '0.0.1'
basePath: /api/v1
paths:
...
  /user/{id}:
    delete:
      operationId: 'delete User'
      description: 'delete User'
      tags:
        - 'user'
      produces:
        - 'application/json'
      parameters:
        - type: integer
          in: path
          name: id
          required: true
      responses:
        200:
          description: 'deletion success'
          schema:
            type: object
            properties:
              message:
                type: string
        401:
          description: 'Unauthorized'
        403:
          description: 'Forbidden'
        404:
          description: 'resource not found'
          schema:
            type: object
            properties:
              message:
                type: string
    put:
      operationId: 'Modify User'
      description: 'Modify User'
      tags:
        - 'user'
      produces:
        - 'application/json'
      parameters:
        - type: integer
          in: path
          name: id
          required: true
        - in: body
          name: reqBody
          schema:
            type: object
            properties:
              name:
                type: string
              org:
                type: string
              role:
                type: string
              email:
                type: string
              phone:
                type: string
      responses:
        200:
          description: 'deletion success'
          schema:
            type: object
            properties:
              message:
                type: string
        401:
          description: 'Unauthorized'
        403:
          description: 'Forbidden'
        404:
          description: 'resource not found'
          schema:
            type: object
            properties:
              message:
                type: string
...